Add files via upload
Food Ordering App Description:-
A food ordering app is a software application that allows users to order food from a restaurant or food delivery service. It typically includes a menu with prices, an ordering system, and a payment system. Some food ordering apps may also include features such as tracking the delivery or pickup status of the order, providing reviews and ratings of restaurants, and offering discounts or promotions.
Modules For this project:-
Importing Json File:
This food ordering app is a command-line application that allows users to order food from a predefined menu. The app is built using Python and imports a JSON file that contains the menu items and their prices. The JSON file can be easily updated without having to modify the code.

Importing Datetime Module:
The datetime module is used to add a timestamp to each order. This allows the app to keep track of when each order was placed.

Importing Json Decoder:
The JSON encoder is used to convert the data into a JSON format that can be easily saved to a file or sent over the internet.